Add files via upload
MVC - Is Divided into Three Main Parts --> Model - View - Controller, The Model Is Where all the functions and methods should be implemented, The View is usually the form or the Html & Css that the user sees and interacts with, and finally The Controller which is the brains of the whole operation; it connects the Model to the View. 
This Method is perfect for creating Web Applications, so you wouldn't have to write all the code in the same file AKA "Spaghetti" Code.

User.php --> Model.
SignUp.php --> View.
UserController.php --> Controller.